+++ app/src/main/java/app/revanced/manager/util/FileFilters.kt
@@ -8,3 +8,8 @@ fun isAllowedApkFile(path: Path): Boolean {
     val extension = path.name.substringAfterLast('.', "").lowercase(Locale.ROOT)
     return extension in APK_FILE_EXTENSIONS
 }
+
+fun isAllowedRvpFile(path: Path): Boolean {
+    val extension = path.name.substringAfterLast('.', "").lowercase(Locale.ROOT)
+    return extension == "rvp"
+}