+++ api/CHANGELOG.md
@@ -7,6 +7,7 @@
 - Added a split APK not support guard for the Rooted mount installer
 - Added additional guards and checks during patch bundle importing using a remote URL
 - Added a "Use device langauge" option that uses the devices set langauge if avaliable, if not it falls back to English
+- Importing patch bundles from a file now shows the progress of the import (similar to how importing remote bundles work)
 
 # Bug fixes
 
@@ -14,6 +15,7 @@
 - Fixed an occasional crash that occurred with some users when opening the app and quickly going to the app selection page and opening a "Show suggested versions" expandable
 - Fixed instability of Rooted mount installer
 - Fixed an issue where the Rooted mount installer would be selectable for users who are non-root
+- Fixed importing patch bundles from storage taking much longer than it should to import (larger patch bundles till take some time)
 
 
 # v1.6.0 (2025-12-17)